requirements.txt
playwright
pytest
toml
python-dotenv